+++ Back-end/src/main/java/com/example/backend/common/auth/AuthUtil.java
@@ -1,0 +1,18 @@
+package com.example.backend.common.auth;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+
+import java.util.UUID;
+
+public class AuthUtil {
+
+    // ✅ 현재 사용자 ID 가져오기
+    public static UUID getCurrentUserId() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new IllegalStateException("인증되지 않은 사용자입니다.");
+        }
+        return UUID.fromString(authentication.getName());
+    }
+}
