+++ Back-end/src/main/java/com/example/backend/user/service/UserService.java
@@ -7,55 +7,26 @@ import com.example.backend.jwt.dto.JwtDto;
 import com.example.backend.user.dto.request.UserRequest;
 import com.example.backend.user.dto.response.UserResponse;
 import com.example.backend.user.entity.User;
+import com.example.backend.user.filter.UserFilter;
 import com.example.backend.user.repository.UserRepository;
+import com.example.backend.common.auth.AuthUtil;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Map;
-import java.util.UUID;
-import java.util.concurrent.ConcurrentHashMap;
 
 @Service
 @RequiredArgsConstructor
 public class UserService {
     private final Logger log = LoggerFactory.getLogger(UserService.class);
+
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
     private final JWTGenerator jwtGenerator;
-
-    // ✅ UserId 구하기
-    private UUID getCurrentUserId() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (authentication == null || !authentication.isAuthenticated()) {
-            throw new IllegalStateException("인증되지 않은 사용자입니다.");
-        }
-        return UUID.fromString(authentication.getName());
-    }
-
-    // ✅ 로그인 실패 기록
-    private final Map<String, LoginAttempt> loginFailures = new ConcurrentHashMap<>();
-    private static final int MAX_ATTEMPTS = 5;
-    private static final long BLOCK_DURATION_MS = 5 * 60 * 1000;
-
-    private static class LoginAttempt {
-        int count;
-        long lastAttemptTime;
-
-        LoginAttempt(int count, long lastAttemptTime) {
-            this.count = count;
-            this.lastAttemptTime = lastAttemptTime;
-        }
-    }
-
-    @Autowired
-    private CartRepository cartRepository;
+    private final CartRepository cartRepository;
+    private final UserFilter userFilter;
 
     // 1️⃣ 회원가입 로직
     @Transactional
@@ -89,7 +60,7 @@ public class UserService {
     // 2️⃣ 유저정보 변경 로직
     @Transactional
     public UserResponse.updateResponse update(UserRequest.updateRequest request) {
-        User user = userRepository.findById(getCurrentUserId())
+        User user = userRepository.findById(AuthUtil.getCurrentUserId())
                 .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
 
         if (request.getEmail() != null) {
@@ -120,7 +91,7 @@ public class UserService {
     @Transactional
     public void delete() {
         userRepository.delete(
-                userRepository.findById(getCurrentUserId())
+                userRepository.findById(AuthUtil.getCurrentUserId())
                         .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."))
         );
     }
@@ -128,7 +99,7 @@ public class UserService {
     // 4️⃣ 유저 정보 조회
     @Transactional
     public UserResponse.InformationResponse myInfo() {
-        User user = userRepository.findById(getCurrentUserId())
+        User user = userRepository.findById(AuthUtil.getCurrentUserId())
                 .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
 
         return UserResponse.InformationResponse.builder()
@@ -144,22 +115,20 @@ public class UserService {
     public UserResponse.loginResponse login(UserRequest.loginRequest request) {
         String email = request.getEmail();
 
-        if (isBlocked(email)) {
-            throw new IllegalArgumentException("로그인 시도가 너무 많습니다. 잠시 후 다시 시도해주세요.");
-        }
-
         User user = userRepository.findByEmail(email)
                 .orElseThrow(() -> {
-                    recordLoginFailure(email);
+                    userFilter.recordLoginFailure(email);
                     return new IllegalArgumentException("이메일이 존재하지 않습니다.");
                 });
 
         if (!passwordEncoder.matches(request.getPassword(), user.getPassword())) {
-            recordLoginFailure(email);
+            userFilter.recordLoginFailure(email);
             throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
         }
 
-        resetLoginFailures(email);
+        // ✅ 성공 시 기록 초기화
+        userFilter.resetLoginFailures(email);
+
         log.info("로그인 성공 - 사용자 이메일 : {}", email);
 
         JwtDto jwtDto = jwtGenerator.generateToken(user);
@@ -167,7 +136,7 @@ public class UserService {
         return UserResponse.loginResponse.builder()
                 .jwtDto(jwtDto)
                 .userNickname(user.getUserNickname())
-                .userProfileImage(user.getUserProfileImage()) // ✅ 바로 반환
+                .userProfileImage(user.getUserProfileImage())
                 .userRole(user.getUserRole())
                 .build();
     }
@@ -183,33 +152,9 @@ public class UserService {
         userRepository.save(user);
     }
 
+    // ✅ 이메일 유효성 체크
     private boolean isValidEmail(String email) {
         String emailRegex = "^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+$";
         return email != null && email.matches(emailRegex);
-    }
-
-    private void recordLoginFailure(String email) {
-        LoginAttempt attempt = loginFailures.getOrDefault(email, new LoginAttempt(0, System.currentTimeMillis()));
-        attempt.count++;
-        attempt.lastAttemptTime = System.currentTimeMillis();
-        loginFailures.put(email, attempt);
-    }
-
-    private void resetLoginFailures(String email) {
-        loginFailures.remove(email);
-    }
-
-    private boolean isBlocked(String email) {
-        LoginAttempt attempt = loginFailures.get(email);
-        if (attempt == null) return false;
-
-        long now = System.currentTimeMillis();
-
-        if (now - attempt.lastAttemptTime > BLOCK_DURATION_MS) {
-            loginFailures.remove(email);
-            return false;
-        }
-
-        return attempt.count >= MAX_ATTEMPTS;
     }
 }
